+++ pages/api/session/logout.ts
@@ -1,7 +1,9 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
 import nc from 'next-connect';
 
+import { cookieName } from 'config/constants';
 import { onError, onNoMatch } from 'lib/middleware';
+import { getCookieDomain } from 'lib/session/getCookieDomain';
 import { withSessionRoute } from 'lib/session/withSession';
 
 const handler = nc<NextApiRequest, NextApiResponse>({ onError, onNoMatch });
@@ -10,6 +12,19 @@ handler.post(logout);
 
 async function logout(req: NextApiRequest, res: NextApiResponse<{ ok: boolean }>) {
   req.session.destroy();
+  const cookieDomain = getCookieDomain();
+
+  if (cookieDomain) {
+    const cookiesHeader = res.getHeader('Set-Cookie') || [];
+    const setCookiesArray = Array.isArray(cookiesHeader) ? cookiesHeader : [];
+
+    res.setHeader('Set-Cookie', [
+      ...setCookiesArray,
+      // remove old domain cookies
+      `${cookieName}=; Max-Age=0; Domain=app.${getCookieDomain()}; Path=/; HttpOnly; SameSite=Strict`
+    ]);
+  }
+
   res.send({ ok: true });
 }
